extract extension to new file
